+++ src/main/java/com/example/backend/Entity/Survey.java
@@ -1,0 +1,6 @@
+package com.example.backend.Entity;
+
+public class Survey {
+
+
+}
