+++ src/main/java/com/example/backend/mapper/UserMapper.java
@@ -1,11 +1,18 @@
 package com.example.backend.mapper;
-import com.example.backend.po.User;
+import com.example.backend.Entity.User;
+import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+
+
+// This is the mapper class for User entity.
 
 @Mapper
 public interface UserMapper {
 
     @Select("select * from user where id = #{id}")
-    public User getUserById(Integer id);
+    User getUserById(Integer id);
+
+    @Insert("insert into user (name, age, gender, phone, password) VALUES (#{name}, #{age}, #{gender}, #{phone}, #{password})")
+    int insertUser(User user);
 }
