+++ src/main/java/com/example/backend/service/UserService.java
@@ -1,6 +1,8 @@
 package com.example.backend.service;
-import com.example.backend.po.User;
+import com.example.backend.Entity.User;
 public interface UserService {
 
     public User findById(Integer id);
+    public int insertUser(User user);
+
 }
