+++ src/main/java/com/example/backend/service/impl/UserServiceImpl.java
@@ -1,10 +1,10 @@
 package com.example.backend.service.impl;
 import com.example.backend.service.UserService;
-import com.example.backend.po.User;
+import com.example.backend.Entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.example.backend.mapper.UserMapper;
-import org.springframework.beans.factory.annotation.Autowired;
+
 @Service
 public class UserServiceImpl implements UserService {
 
@@ -15,4 +15,11 @@ public class UserServiceImpl implements UserService {
     public User findById(Integer id) {
         return userMapper.getUserById(id);
     }
+
+    @Override
+    public int insertUser(User user) {
+        return userMapper.insertUser(user);
+    }
+
+
 }
