+++ src/main/resources/application.yml
@@ -5,7 +5,7 @@ spring:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://localhost:3306/mybatis
     username: root
-    password: helloworld
+    password: 123456
 
 server:
   port: 8080
